refactor: Refactor appointment creation and calendar functionality
